refactor: cms pages
