+++ packages/scaffolding-cli/templates/src/ssr/components/MainFeaturedPost/index.tsx
@@ -74,7 +74,10 @@ export default function MainFeaturedPost(props) {
                             <Typography variant="h5" color="inherit" paragraph>
                                 {excerpt}
                             </Typography>
-                            <Link as={`/posts/${slug}`} href="/posts/[...all]">
+                            <Link
+                                as={`/cms/posts/${slug}`}
+                                href="/cms/posts/[...all]"
+                            >
                                 <a className={classes.readMoreButton}>
                                     Read More
                                 </a>
